move diffable into snapshot
