+++ lib/active_snapshot/models/snapshot.rb
@@ -2,8 +2,6 @@ module ActiveSnapshot
   class Snapshot < ActiveRecord::Base
     self.table_name = "snapshots"
 
-    include ActiveSnapshot::Diffable
-
     if defined?(ProtectedAttributes)
       attr_accessible :item_id, :item_type, :identifier, :user_id, :user_type
     end
@@ -17,25 +15,114 @@ module ActiveSnapshot
     validates :identifier, uniqueness: { scope: [:item_id, :item_type], allow_nil: true}
     validates :user_type, presence: true, if: :user_id
 
-    def self.build_snapshot(resource, identifier: nil, user: nil, metadata: nil)
-      snapshot = resource.snapshots.build({
-        identifier: identifier,
-        user_id: (user.id if user),
-        user_type: (user.class.name if user),
-        metadata: (metadata || {}),
-      })
+    class << self
+      def build_snapshot(resource, identifier: nil, user: nil, metadata: nil)
+        snapshot = resource.snapshots.build({
+          identifier: identifier,
+          user_id: (user.id if user),
+          user_type: (user.class.name if user),
+          metadata: (metadata || {}),
+        })
 
-      snapshot.build_snapshot_item(resource)
+        snapshot.build_snapshot_item(resource)
 
-      snapshot_children = resource.children_to_snapshot
+        snapshot_children = resource.children_to_snapshot
 
-      snapshot_children&.each do |child_group_name, h|
-        h[:records].each do |child_item|
-          snapshot.build_snapshot_item(child_item, child_group_name: child_group_name)
+        snapshot_children&.each do |child_group_name, h|
+          h[:records].each do |child_item|
+            snapshot.build_snapshot_item(child_item, child_group_name: child_group_name)
+          end
         end
+
+        snapshot
       end
 
-      snapshot
+      def diff(from, to)
+        raise ArgumentError, "'from' must be an ActiveSnapshot::Snapshot" unless from.is_a?(Snapshot)
+
+        to_item_id, to_item_type = to.is_a?(Snapshot) ? [to.item_id, to.item_type] : [to.id, to.class.name]
+
+        if from.item_id != to_item_id || from.item_type != to_item_type
+          raise ArgumentError, "Both records must reference the same item"
+        end
+
+        if to.is_a?(Snapshot) && from.created_at > to.created_at
+          raise ArgumentError, "'to' must be a newer snapshot than 'from'"
+        end
+
+        from_snapshot = from
+        to_snapshot = to.is_a?(Snapshot) ? to : build_snapshot(to)
+
+        from_snapshot_items = from_snapshot.snapshot_items
+        to_snapshot_items = to_snapshot.snapshot_items
+
+        diffs = []
+
+        from_snapshot_items.each do |from_snapshot_item|
+          to_snapshot_item = to_snapshot_items.find do |item|
+            item.item_id == from_snapshot_item.item_id && item.item_type == from_snapshot_item.item_type
+          end
+
+          if to_snapshot_item.nil?
+            diffs << {
+              action: :destroy,
+              item_id: from_snapshot_item.item_id,
+              item_type: from_snapshot_item.item_type,
+              changes: snapshot_item_changes(from_snapshot_item, nil)
+            }
+          else
+            changes = snapshot_item_changes(from_snapshot_item, to_snapshot_item)
+
+            next if changes.empty?
+
+            diffs << {
+              action: :update,
+              item_id: from_snapshot_item.item_id,
+              item_type: from_snapshot_item.item_type,
+              changes: changes
+            }
+          end
+        end
+
+        to_snapshot_items.each do |to_snapshot_item|
+          from_snapshot_item = from_snapshot_items.find do |item|
+            item.item_id == to_snapshot_item.item_id && item.item_type == to_snapshot_item.item_type
+          end
+
+          next if from_snapshot_item.present?
+
+          diffs << {
+            action: :create,
+            item_id: to_snapshot_item.item_id,
+            item_type: to_snapshot_item.item_type,
+            changes: snapshot_item_changes(nil, to_snapshot_item)
+          }
+        end
+
+        diffs
+      end
+
+      private
+
+      def snapshot_item_changes(from, to)
+        from_object = from ? from.object : {}
+        to_object = to ? to.object : {}
+
+        keys = (from_object.keys + to_object.keys).uniq
+
+        changes = {}
+
+        keys.each do |key|
+          from_value = from_object[key]
+          to_value = to_object[key]
+
+          next if to_value == from_value
+
+          changes[key.to_sym] = [from_value, to_value]
+        end
+
+        changes
+      end
     end
 
     def metadata
